tiny-daily-assistants/tourist-guide/requirements.txt
openai>=1.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
pydub>=0.25.1
pypdf>=4.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
tiktoken>=0.5.2